pyproject.toml
[project]
name = "bibtex-cleaner"
version = "0.1.1"
description = "LLM-powered BibTeX cleaner"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    {name = "Mengyu Ye", email = "ye.mengyu.s1@dc.tohoku.ac.jp"},
    {name = "Ryosuke Takahashi", email = "takahashi.ryosuke.r7@dc.tohoku.ac.jp"},
    {name = "Taïga Gonçalves", email = "goncalves.taiga.teo.t6@dc.tohoku.ac.jp"},
    {name = "Kai Kamijo", email = "kamijou.kai.p5@dc.tohoku.ac.jp"},
]
license = {file = "LICENSE"}
keywords = ["bibtex", "citation", "langchain", "cli", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Topic :: Text Processing :: Markup :: LaTeX",
]
dependencies = [
    "bs4>=0.0.2",
    "fire>=0.7.1",
    "langchain-community>=0.3.27",
    "langchain-openai>=0.3.0",
    "langchain-core>=0.3.0",
    "lxml>=6.0.0",
    "openai>=1.93.0",
    "python-dotenv>=1.1.1",
    "requests>=2.31.0",
    "tiktoken>=0.8.0,<0.11.0",
    "tqdm>=4.67.1",
    "pyyaml>=6.0"
]

[project.urls]
Homepage = "https://github.com/tohoku-nlp/bibtex-cleaning-agent"
Repository = "https://github.com/tohoku-nlp/bibtex-cleaning-agent"
Issues = "https://github.com/tohoku-nlp/bibtex-cleaning-agent/issues"

[project.scripts]
bibtex-cleaner = "bibtex_cleaner.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
sources = ["src"]
[tool.hatch.build.targets.wheel]
packages = ["bibtex_cleaner"]

[dependency-groups]
dev = [
    "ipython>=9.4.0",
    "pytest>=8.4.1",
]
